+++ screen-orientation/Package.swift
@@ -3,21 +3,21 @@ import PackageDescription
 
 let package = Package(
     name: "CapacitorScreenOrientation",
-    platforms: [.iOS(.v13)],
+    platforms: [.iOS(.v14)],
     products: [
         .library(
             name: "CapacitorScreenOrientation",
             targets: ["ScreenOrientationPlugin"])
     ],
     dependencies: [
-        .package(url: "https://github.com/ionic-team/capacitor-spm.git", branch: "main")
+        .package(url: "https://github.com/ionic-team/capacitor-swift-pm.git", branch: "main")
     ],
     targets: [
         .target(
             name: "ScreenOrientationPlugin",
             dependencies: [
-                .product(name: "Capacitor", package: "capacitor-spm"),
-                .product(name: "Cordova", package: "capacitor-spm")
+                .product(name: "Capacitor", package: "capacitor-swift-pm"),
+                .product(name: "Cordova", package: "capacitor-swift-pm")
             ],
             path: "ios/Sources/ScreenOrientationPlugin"),
         .testTarget(