+++ keyboard/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.2](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@6.0.1...@capacitor/keyboard@6.0.2) (2024-08-08)
+
+**Note:** Version bump only for package @capacitor/keyboard
+
+## [6.0.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@6.0.0...@capacitor/keyboard@6.0.1) (2024-06-13)
+
+### Bug Fixes
+
+- **ios:** apply filter to fetch correct scene object ([#2102](https://github.com/ionic-team/capacitor-plugins/issues/2102)) ([acd334d](https://github.com/ionic-team/capacitor-plugins/commit/acd334dae7fc7545446cafc8fbb469200f7831f8))
+
+# [6.0.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@6.0.0-rc.1...@capacitor/keyboard@6.0.0) (2024-04-15)
+
+**Note:** Version bump only for package @capacitor/keyboard
+
+# [6.0.0-rc.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@6.0.0-rc.0...@capacitor/keyboard@6.0.0-rc.1) (2024-03-25)
+
+**Note:** Version bump only for package @capacitor/keyboard
+
 # [6.0.0-rc.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/keyboard@6.0.0-beta.1...@capacitor/keyboard@6.0.0-rc.0) (2024-02-07)
 
 **Note:** Version bump only for package @capacitor/keyboard